requirements.txt
crewai
crewai-tools
load_dotenv
langchain-huggingface
langchain-groq
langchain-openai